fix: move multi-handling to dedicated modules for cleaner argument-validation (#320, #334)
