+++ docs/source/modules/rule_multi.rst
@@ -68,12 +68,12 @@ Basics
           firewall: 'opnsense.template.opnsense.oxl.app'
           api_credential_file: '/home/guy/.secret/opn.key'
 
-        oxlorg.opnsense.rule:
+        oxlorg.opnsense.rule_multi:
           match_fields: ['description']
 
       tasks:
         - name: Changing
-          oxlorg.opnsense.rule:
+          oxlorg.opnsense.rule_multi:
             rules:
               - name: 'test1'
                 source_net: '192.168.1.0/24'
@@ -118,7 +118,7 @@ Basics
             var: existing_entries.data
 
         - name: Purging all non-configured rules
-          oxlorg.opnsense.rule:
+          oxlorg.opnsense.rule_multi:
             rules: {...}
 
             match_fields: ['description']
@@ -127,7 +127,7 @@ Basics
               # action: 'disable'  # default = delete
 
         - name: Purging allow-rules on interface opt2 that use IPv4
-          oxlorg.opnsense.rule:
+          oxlorg.opnsense.rule_multi:
             multi_control:
               purge_all: true
               filters:  # filtering rules to purge by rule-parameters
@@ -145,7 +145,7 @@ You can also override all rule parameters as needed.
 .. code-block:: yaml
 
     - name: Changing
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules: {...}
 
         multi_control:
@@ -164,7 +164,7 @@ To simplify the modules usage and config - you can also use shorter parameter al
 .. code-block:: yaml
 
     - name: Changing
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'test1'
             src: 'ALIAS_URLTABLE_TOR_EXIT_NODES'
