+++ plugins/modules/bind_record_multi.py
@@ -12,10 +12,16 @@ from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.handler impor
     module_dependency_error, MODULE_EXCEPTIONS
 
 try:
-    from ansible_collections.oxlorg.opnsense.plugins.module_utils.defaults.legacy_multi import \
-        FAIL_MOD_ARG_MULTI, INFO_MOD_ARG
+    from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.wrapper import \
+        module_multi_wrapper
+    from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.multi import \
+        build_multi_mod_args, MultiModuleCallbacks
     from ansible_collections.oxlorg.opnsense.plugins.module_utils.defaults.main import \
-        OPN_MOD_ARGS, RELOAD_MOD_ARG, STATE_MOD_ARG
+        OPN_MOD_ARGS, RELOAD_MOD_ARG
+    from ansible_collections.oxlorg.opnsense.plugins.module_utils.defaults.bind import \
+        BIND_REC_MOD_ARGS, BIND_REC_MATCH_FIELDS
+    from ansible_collections.oxlorg.opnsense.plugins.module_utils.main.bind_record import \
+        Record
 
 except MODULE_EXCEPTIONS:
     module_dependency_error()
@@ -25,28 +31,60 @@ except MODULE_EXCEPTIONS:
 # EXAMPLES = 'https://ansible-opnsense.oxl.app/modules/bind.html'
 
 
+class MultiCallbacks(MultiModuleCallbacks):
+    @staticmethod
+    def get_existing(meta_entry: Record) -> dict:
+        meta_entry.search_call_domains()
+        return {
+            'main': meta_entry.get_existing(),
+            'domains': meta_entry.existing_domains,
+        }
+
+    @staticmethod
+    def set_existing(entry: Record, cache: dict):
+        entry.existing_entries = cache['main']
+        entry.existing_domains = cache['domains']
+
+
 def run_module():
-    module_args = dict(
-        records=dict(type='dict', required=True),
-        match_fields=dict(
-            type='list', required=False, elements='str',
-            description='Fields that are used to match configured records with the running config - '
-                        "if any of those fields are changed, the module will think it's a new entry",
-            choices=['domain', 'name', 'type', 'value'],
-            default=['domain', 'name', 'type'],
-        ),
-        **FAIL_MOD_ARG_MULTI,
-        **STATE_MOD_ARG,
-        **INFO_MOD_ARG,
-        **OPN_MOD_ARGS,
-        **RELOAD_MOD_ARG,
+    entry_multi_args = build_multi_mod_args(
+        mod_args=BIND_REC_MOD_ARGS,
+        aliases=['records'],
+        not_required=['domain'],  # overrides could be used to define the domain
     )
 
-    AnsibleModule(
+    module_args = dict(
+        **entry_multi_args,
+        **RELOAD_MOD_ARG,
+        **OPN_MOD_ARGS,
+        **BIND_REC_MATCH_FIELDS,
+    )
+
+    module = AnsibleModule(
         argument_spec=module_args,
         supports_check_mode=True,
-    ).fail_json('This module was deprecated in favor of: https://ansible-opnsense.oxl.app/modules/1_multi.html')
+        required_one_of=[
+            ('multi', 'multi_purge', 'multi_control.purge_all'),
+        ],
+    )
 
+    result = dict(
+        changed=False,
+        diff={
+            'before': {},
+            'after': {},
+        }
+    )
+
+    module_multi_wrapper(
+        module=module,
+        result=result,
+        obj=Record,
+        kind='record',
+        entry_args=entry_multi_args,
+        callbacks=MultiCallbacks(),
+    )
+    module.exit_json(**result)
 
 
 def main():
