+++ plugins/modules/rule.py
@@ -14,9 +14,7 @@ from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.handler impor
 
 try:
     from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.wrapper import \
-        module_wrapper, is_multi_module_call, module_multi_wrapper
-    from ansible_collections.oxlorg.opnsense.plugins.module_utils.base.multi import \
-        build_multi_mod_args
+        module_wrapper
     from ansible_collections.oxlorg.opnsense.plugins.module_utils.defaults.main import \
         OPN_MOD_ARGS, RELOAD_MOD_ARG
     from ansible_collections.oxlorg.opnsense.plugins.module_utils.defaults.rule import RULE_MOD_ARGS
@@ -31,15 +29,8 @@ except MODULE_EXCEPTIONS:
 
 
 def run_module():
-    entry_args = RULE_MOD_ARGS
-    entry_multi_args = build_multi_mod_args(
-        mod_args=entry_args,
-        aliases=['rules'],
-    )
-
     module_args = dict(
-        **entry_args,
-        **entry_multi_args,
+        **RULE_MOD_ARGS,
         **OPN_MOD_ARGS,
         **RELOAD_MOD_ARG,
     )
@@ -47,12 +38,6 @@ def run_module():
     module = AnsibleModule(
         argument_spec=module_args,
         supports_check_mode=True,
-        mutually_exclusive=[
-            ('description', 'multi'), ('description', 'multi_purge'), ('description', 'multi_control.purge_all')
-        ],
-        required_one_of=[
-            ('description', 'multi', 'multi_purge', 'multi_control.purge_all'),
-        ],
     )
 
     result = dict(
@@ -63,18 +48,7 @@ def run_module():
         },
     )
 
-    if is_multi_module_call(module):
-        module_multi_wrapper(
-            module=module,
-            result=result,
-            obj=Rule,
-            kind='rule',
-            entry_args=entry_multi_args,
-        )
-
-    else:
-        module_wrapper(Rule(module=module, result=result))
-
+    module_wrapper(Rule(module=module, result=result))
     module.exit_json(**result)
