+++ tests/rule_multi.yml
@@ -12,7 +12,7 @@
     oxlorg.opnsense.list:
       target: 'rule'
 
-    oxlorg.opnsense.rule:
+    oxlorg.opnsense.rule_multi:
       match_fields: ['description']
       multi_control:
         fail_verify: true
@@ -23,7 +23,7 @@
 
   tasks:
     - name: Removing - do not exist
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_1'
           - name: 'ANSIBLE_TEST_2_2'
@@ -39,7 +39,7 @@
         opn1.changed
 
     - name: Adding
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_1'
             source_net: '192.168.1.0/24'
@@ -62,7 +62,7 @@
         not opn2.changed
 
     - name: Changing
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_1'
             source_net: '192.168.1.0/24'
@@ -87,7 +87,7 @@
         not opn3.changed
 
     - name: Changing - not changed
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_1'
             source_net: '192.168.1.0/24'
@@ -112,7 +112,7 @@
         opn7.changed
 
     - name: Adding 2
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_4'
             interface: 'opt1'
@@ -124,7 +124,7 @@
         not opn8.changed
 
     - name: Changing 2 - not changed
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_4'
             interface: 'opt1'
@@ -136,7 +136,7 @@
         opn9.changed
 
     - name: Fail on server-side validation
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_1'
           - name: 'ANSIBLE_TEST_2_2'
@@ -147,7 +147,7 @@
       when: not ansible_check_mode
 
     - name: Adding 3 - Advanced Mode
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_5'
             allow_opts: true
@@ -173,7 +173,7 @@
         not opn10.changed
 
     - name: Changing 3 - Advanced Mode - not changed
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_5'
             allow_opts: true
@@ -201,7 +201,7 @@
 
     # noting to validate..
     # - name: Fail on client-side validation
-    #   oxlorg.opnsense.rule:
+    #   oxlorg.opnsense.rule_multi:
     #     rules:
     #       - name: 'ANSIBLE_TEST_2_1'
     #       - name: 'ANSIBLE_TEST_2_2'
@@ -211,7 +211,7 @@
     #   failed_when: not opn6.failed
     #
     # - name: Don't fail on client-side validation
-    #   oxlorg.opnsense.rule:
+    #   oxlorg.opnsense.rule_multi:
     #     rules:
     #       - name: 'ANSIBLE_TEST_2_1'
     #       - name: 'ANSIBLE_TEST_2_2'
@@ -231,7 +231,7 @@
       when: not ansible_check_mode
 
     - name: Removing
-      oxlorg.opnsense.rule:
+      oxlorg.opnsense.rule_multi:
         rules:
           - name: 'ANSIBLE_TEST_2_1'
           - name: 'ANSIBLE_TEST_2_2'
